Refactor folder structure to follow feature-based approach
