--- src/models/userModel.js
@@ -1,20 +1,0 @@
-import mongoose from 'mongoose';
-
-const userSchema = new mongoose.Schema({
-  _id: {
-    type: String,
-    required: true,
-  },
-  avatarURL: {
-    type: String,
-    // TODO: Rework with image storage
-    default: 'https://i.insider.com/602ee9ced3ad27001837f2ac?width=700',
-    required: true,
-  },
-  habits: {
-    type: [],
-    default: [],
-  },
-});
-
-export default mongoose.model('User', userSchema);
