+++ src/server.js
@@ -2,9 +2,9 @@ import stoppable from 'stoppable';
 
 import app from './app.js';
 import config from './config.js';
-import logger from './utils/logger.js';
-import { connectDB } from './utils/dbConnection.js';
-import { hydrateCognitoJwks } from './utils/auth.js';
+import logger from './shared/utils/logger.js';
+import { connectDB } from './core/dbConnection.js';
+import { hydrateCognitoJwks } from './core/auth.js';
 
 let server = null;
