+++ src/shared/models/event.model.js
